PR updates

+++ lib/src/qr_codes/qr_code_image_scanning.dart
@@ -4,7 +4,12 @@ import 'package:zxing2/qrcode.dart';
 /// QR code extensions on [Image].
 extension ImageQrScanning on Image {
   /// Attempts to find and decode a QR code within this [Image].
-  QrCode readQrCode() {
+  ///
+  /// If there are multiple QR codes in the image, it returns the first one it finds
+  /// from top to bottom.
+  ///
+  /// If no QR code is found, it returns `null`.
+  QrCode? readQrCode() {
     return _readQrCode(this);
   }
 
@@ -15,41 +20,42 @@ extension ImageQrScanning on Image {
     Image image = this;
 
     // Keep looking for QR codes until we can't find anymore.
-    while (true) {
-      try {
-        final qrcode = _readQrCode(image);
-        qrCodes.add(qrcode);
-
-        /// Extract the rectangles from the edge points.
-        final bottomLeftRectangle = qrcode.resultPoints[0];
-        final topLeftRectangle = qrcode.resultPoints[1];
-        final topRightRectangle = qrcode.resultPoints[2];
-
-        final top = topRightRectangle.y.truncate();
-        final left = topLeftRectangle.x.truncate();
-        final right = topRightRectangle.x.truncate();
-        final bottom = bottomLeftRectangle.y.truncate();
-
-        // Fill the QRCode region with black color to try to read other QR Codes.
-        image = fillRect(
-          image,
-          x1: left,
-          y1: top,
-          x2: right,
-          y2: bottom,
-          color: ColorRgb8(0, 0, 0),
-        );
-      } on CouldNotReadQrCodeException catch (_) {
-        // We didn't find any QR codes in the image. Stop looking.
-        break;
+    bool keepSearching = true;
+    while (keepSearching) {
+      final qrcode = _readQrCode(image);
+      if (qrcode == null) {
+        keepSearching = false;
+        continue;
       }
+
+      qrCodes.add(qrcode);
+
+      /// Extract the rectangles from the edge points.
+      final bottomLeftRectangle = qrcode.resultPoints[0];
+      final topLeftRectangle = qrcode.resultPoints[1];
+      final topRightRectangle = qrcode.resultPoints[2];
+
+      final top = topRightRectangle.y.truncate();
+      final left = topLeftRectangle.x.truncate();
+      final right = topRightRectangle.x.truncate();
+      final bottom = bottomLeftRectangle.y.truncate();
+
+      // Fill the QRCode region with black color to try to read other QR Codes.
+      image = fillRect(
+        image,
+        x1: left,
+        y1: top,
+        x2: right,
+        y2: bottom,
+        color: ColorRgb8(0, 0, 0),
+      );
     }
 
     return qrCodes;
   }
 }
 
-QrCode _readQrCode(Image image) {
+QrCode? _readQrCode(Image image) {
   final rgbLuminanceSource = RGBLuminanceSource(
     image.width,
     image.height,
@@ -78,18 +84,10 @@ QrCode _readQrCode(Image image) {
     );
     final binaryBitmap = BinaryBitmap(binarizer);
     return QRCodeReader().decode(binaryBitmap);
-  } on ReaderException catch (exception) {
-    throw CouldNotReadQrCodeException(exception);
+  } on ReaderException {
+    // We didn't find a QR code in the image.
+    return null;
   }
 }
 
 typedef QrCode = Result;
-
-class CouldNotReadQrCodeException implements Exception {
-  const CouldNotReadQrCodeException([this.rootCause]);
-
-  final Exception? rootCause;
-
-  @override
-  String toString() => "CouldNotReadQrCodeException";
-}
